requirements/base.txt
# Core Django packages
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0
django-filter>=23.0
python-decouple>=3.8
Pillow>=9.5.0

# Development utilities
django-extensions>=3.2.0
